requirements.txt
beautifulsoup4==4.12.2
lxml==4.9.3
Markdown==3.4.4
Pillow==9.5.0
prs==0.0.7
python-pptx==0.6.22
soupsieve==2.5
XlsxWriter==3.1.5
